resolvendo problema de exlcusão do meu bacno de dados

+++ routes/guestsRoutes.js
@@ -95,6 +95,10 @@ router.put('/editGuest/:id', async (req, res) => {
 router.delete('/deleteGuest/:id', async (req, res) => {
   const { id } = req.params;
 
+  if (!existingGuest) {
+    return res.status(404).send('O hóspede que você está tentando excluir não existe.');
+  }
+
   try {
     await prisma.guest.delete({
       where: {
